Modified - move store model
