+++ lib/features/store_feature/business/entities/orderItem.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/foundation.dart';
 
-import '../models/color_code.dart';
+import '../../../../models/color_code.dart';
 
 class OrderItem with ChangeNotifier {
   final int id;
