+++ lib/features/store_feature/business/entities/product_cart.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 
-import '../models/color_code_card.dart';
-import 'brandc.dart';
+import '../../../../models/color_code_card.dart';
+import '../../../../models/brandc.dart';
 
 class ProductCart with ChangeNotifier {
   final int id;
