+++ lib/features/store_feature/business/entities/product_main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 
-import '../models/product.dart';
-import '../models/search_detail.dart';
+import 'product.dart';
+import '../../../../models/search_detail.dart';
 
 class ProductMain with ChangeNotifier {
   final SearchDetail productsDetail;
