+++ lib/features/store_feature/presentation/providers/orders.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:http/http.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:tamizshahr/models/order.dart';
-import 'package:tamizshahr/models/order_main.dart';
+import 'package:tamizshahr/features/store_feature/business/entities/order_main.dart';
 import 'package:tamizshahr/models/search_detail.dart';
 
 import '../../../../core/constants/urls.dart';
