+++ lib/features/store_feature/presentation/screens/order_view_screen.dart
@@ -7,7 +7,7 @@ import 'package:url_launcher/url_launcher.dart';
 import '../../../../core/theme/app_theme.dart';
 import '../../../../models/color_code.dart';
 import '../../../../models/gallery.dart';
-import '../../../../models/order_details.dart';
+import '../../business/entities/order_details.dart';
 import '../../../customer_feature/presentation/providers/customer_info.dart';
 import '../../../../core/logic/en_to_ar_number_convertor.dart';
 import '../../../../core/widgets/main_drawer.dart';
