+++ lib/widgets/custom_dialog_select_color.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart' as intl;
 
-import '../models/product.dart';
+import '../features/store_feature/business/entities/product.dart';
 import '../core/theme/app_theme.dart';
 
 class CustomDialogSelectColor extends StatefulWidget {
