+++ lib/widgets/product_item_product_screeen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart' as intl;
 import 'package:provider/provider.dart';
 
-import '../models/product.dart';
+import '../features/store_feature/business/entities/product.dart';
 import '../features/store_feature/presentation/providers/Products.dart';
 import '../core/theme/app_theme.dart';
 import '../features/store_feature/presentation/screens/product_detail_screen.dart';
